serving_patterns/synchronous_pattern/imagenet_inception_v3/requirements.txt
importlib-metadata>=1.7.0
joblib>=0.15.1
numpy>=1.18.5
Pillow>=8.3.2
psutil==5.7.0
pydantic>=1.8.2
PyYAML>=5.3.1
tensorflow-serving-api>=2.5.2
tensorflow-hub>=0.10.0
typing>=3.7.4.1
grpcio==1.34.0
